Move Pool and FastPass out of signals/service into their own modules
service.rs held three independent types in one file. The text-accumulation
pool and the non-OCR fast-pass gather are self-contained concerns, so they
move to sibling modules; service.rs keeps the orchestration (the service
handle, run/stream extraction, and the diagnostic dump). Pure relocation —
no signatures or behavior change. Pool/FastPass state used across the module
boundary becomes pub(super); everything else stays private.
