+++ bae-core/src/signals/fast_pass.rs
@@ -1,0 +1,339 @@
+//! The fast pass: every non-OCR signal a folder yields, gathered in one
+//! blocking hop. [`gather_non_ocr_sources`] scans the folder and reads the
+//! disc ID, CUE-CATALOG barcodes, and classified-text source lines + brackets,
+//! plus the artwork paths handed to the later OCR phase. The result is the
+//! [`FastPass`] the service emits as its first `Signals` snapshot.
+
+use crate::identify::candidate_text::{
+    extract_folder_brackets, parse_filename_stem, strip_path_component, Source, SourcedLine,
+};
+use crate::import::discid::compute_discid_from_categorized;
+use crate::import::folder_scanner::{self, AudioContent, CategorizedFiles};
+use crate::signals::{DiscIdSignal, SignalOrigin, SourcedValue};
+use std::collections::HashSet;
+use std::path::{Path, PathBuf};
+use tracing::{debug, warn};
+
+/// Maximum size read from a single `.txt` file. Caps pathological
+/// inputs (e.g. a 10 MB booklet transcription) without blowing memory.
+const MAX_TEXT_FILE_BYTES: u64 = 100 * 1024;
+
+/// Result of the fast-pass gather: every non-OCR signal a folder yields — the
+/// disc ID, CUE-CATALOG barcodes, classified-text source lines + brackets —
+/// plus the artwork paths for the OCR phase.
+pub(super) struct FastPass {
+    pub(super) lines: Vec<SourcedLine>,
+    pub(super) bracket_catalogs: Vec<String>,
+    pub(super) artwork_paths: Vec<PathBuf>,
+    pub(super) disc_id: DiscIdSignal,
+    pub(super) cue_barcodes: Vec<SourcedValue>,
+}
+
+impl FastPass {
+    /// Empty pass — no disc ID, no sources. Used when the folder scan fails.
+    pub(super) fn empty() -> Self {
+        Self {
+            lines: Vec::new(),
+            bracket_catalogs: Vec::new(),
+            artwork_paths: Vec::new(),
+            disc_id: DiscIdSignal::Absent { track_count: 0 },
+            cue_barcodes: Vec::new(),
+        }
+    }
+}
+
+/// CUE `CATALOG` payloads (the disc's UPC/EAN) from a folder's parsed sheets,
+/// paired and unpaired, deduped — barcode-lookup inputs, not catalog-string
+/// filter values. Each carries `SignalOrigin::CueSheet`.
+fn cue_barcodes(categorized: &CategorizedFiles) -> Vec<SourcedValue> {
+    let mut out: Vec<SourcedValue> = Vec::new();
+    let mut seen: HashSet<String> = HashSet::new();
+    let mut consider = |catalog: &Option<String>| {
+        if let Some(value) = catalog {
+            let value = value.trim();
+            if !value.is_empty() && seen.insert(value.to_string()) {
+                out.push(SourcedValue::new(value.to_string(), SignalOrigin::CueSheet));
+            }
+        }
+    };
+    if let AudioContent::CueFlacPairs { pairs, .. } = &categorized.audio {
+        for pair in pairs {
+            if let Some(sheet) = &pair.cue_sheet {
+                consider(&sheet.catalog);
+            }
+        }
+    }
+    for (_, sheet) in &categorized.unpaired_cue_sheets {
+        consider(&sheet.catalog);
+    }
+    out
+}
+
+/// Synchronously enumerate and read all non-OCR sources. Runs on the
+/// blocking pool via `spawn_blocking`. Any failure surfaces as missing data
+/// for that source — it never aborts extraction.
+pub(super) fn gather_non_ocr_sources(folder: &Path) -> FastPass {
+    let mut pass = FastPass::empty();
+
+    // Folder path components: the candidate folder name + its immediate
+    // parent (the plan's "last 2 path components" heuristic). Strip year /
+    // track-number prefixes and trailing brackets.
+    let folder_name = folder
+        .file_name()
+        .and_then(|n| n.to_str())
+        .unwrap_or_default()
+        .to_string();
+    let parent_name = folder
+        .parent()
+        .and_then(|p| p.file_name())
+        .and_then(|n| n.to_str())
+        .unwrap_or_default()
+        .to_string();
+
+    for raw in [&parent_name, &folder_name] {
+        if raw.is_empty() {
+            continue;
+        }
+        if let Some(stripped) = strip_path_component(raw) {
+            pass.lines.push(SourcedLine {
+                source: Source::PathComponent,
+                text: stripped,
+            });
+        }
+        for bracket in extract_folder_brackets(raw) {
+            pass.bracket_catalogs.push(bracket);
+        }
+    }
+
+    // Enumerate files via the existing folder scanner. Failure here means
+    // no audio was detected — we can still pull filenames / brackets from
+    // the folder name, so we already captured that above.
+    let categorized = match folder_scanner::collect_release_candidate_files(folder) {
+        Ok(c) => c,
+        Err(e) => {
+            debug!(
+                "signals: folder scan failed for {:?}: {e}; using folder-name signals only",
+                folder,
+            );
+            return pass;
+        }
+    };
+
+    // Disc ID from LOG/CUE and CUE-CATALOG barcodes — derived from the same
+    // parsed scan, no re-read.
+    let track_count = categorized.audio.track_count().unwrap_or(0);
+    pass.disc_id = match compute_discid_from_categorized(&categorized) {
+        Some(disc_id) => DiscIdSignal::Computed {
+            disc_id,
+            track_count,
+        },
+        None => DiscIdSignal::Absent { track_count },
+    };
+    pass.cue_barcodes = cue_barcodes(&categorized);
+
+    // Filenames — image + document only. Audio filenames are overwhelmingly
+    // track titles, which are the wrong pool for Artist / Album autocomplete.
+    // Generic names (`cover`, `booklet-01`, etc.) get rejected by
+    // `parse_filename_stem`.
+    for p in enumerate_filename_inputs(&categorized) {
+        for part in parse_filename_stem(&p) {
+            pass.lines.push(SourcedLine {
+                source: Source::FilenameGeneric(p.clone()),
+                text: part,
+            });
+        }
+    }
+
+    // CUE files — harvest PERFORMER / TITLE from the sheets the folder scan
+    // already parsed: paired CUEs carry their sheet, multi-FILE and aggregate
+    // CUEs land in `unpaired_cue_sheets`. No re-reading, no second parser.
+    if let AudioContent::CueFlacPairs { pairs, .. } = &categorized.audio {
+        for pair in pairs {
+            if let Some(sheet) = &pair.cue_sheet {
+                for name in cue_sheet_names(sheet) {
+                    pass.lines.push(SourcedLine {
+                        source: Source::CueField,
+                        text: name,
+                    });
+                }
+            }
+        }
+    }
+    for (_, sheet) in &categorized.unpaired_cue_sheets {
+        for name in cue_sheet_names(sheet) {
+            pass.lines.push(SourcedLine {
+                source: Source::CueField,
+                text: name,
+            });
+        }
+    }
+
+    // Text files — one line per `\n`, treated like OCR input.
+    for doc in text_file_paths(&categorized) {
+        if let Some(text) = read_capped_text(&doc) {
+            for line in text.lines() {
+                let trimmed = line.trim();
+                if !trimmed.is_empty() {
+                    pass.lines.push(SourcedLine {
+                        source: Source::TextFile(doc.clone()),
+                        text: trimmed.to_string(),
+                    });
+                }
+            }
+        }
+    }
+
+    // Artwork paths for the OCR pass.
+    pass.artwork_paths = categorized.artwork.into_iter().map(|f| f.path).collect();
+
+    pass
+}
+
+/// Non-audio filename inputs the classifier should see.
+///
+/// Audio filenames are excluded: their stems are almost always track
+/// titles, which belong in a track-title pool (not surfaced here), not the
+/// Artist / Album autocomplete pool. CUE files are excluded for the same
+/// reason — their `PERFORMER` / `TITLE` values are harvested as `CueField`
+/// lines, while the filename stem (`Album.cue` → `Album`) would only
+/// duplicate path-component signal at lower weight.
+fn enumerate_filename_inputs(categorized: &CategorizedFiles) -> Vec<PathBuf> {
+    let mut out: Vec<PathBuf> = Vec::new();
+    for f in &categorized.artwork {
+        out.push(f.path.clone());
+    }
+    for f in &categorized.documents {
+        out.push(f.path.clone());
+    }
+    out
+}
+
+/// Album- and track-level PERFORMER / TITLE values from a parsed CUE sheet,
+/// deduped — the name tokens for the Artist / Album autocomplete pool. The
+/// folder scanner parses every CUE once; this reads that result rather than
+/// re-scanning the file.
+fn cue_sheet_names(sheet: &crate::cue_flac::CueSheet) -> Vec<String> {
+    let mut out = Vec::new();
+    let mut seen = HashSet::new();
+    let mut push = |value: &Option<String>| {
+        if let Some(s) = value {
+            let s = s.trim();
+            if !s.is_empty() && seen.insert(s.to_string()) {
+                out.push(s.to_string());
+            }
+        }
+    };
+    push(&sheet.performer);
+    push(&sheet.title);
+    for track in &sheet.tracks {
+        push(&track.performer);
+        push(&track.title);
+    }
+    out
+}
+
+/// Returns `.txt` files from the documents list. Excludes `.log`
+/// (rip-technical data with no artist/album content) and `.cue` (handled
+/// separately).
+fn text_file_paths(categorized: &CategorizedFiles) -> Vec<PathBuf> {
+    categorized
+        .documents
+        .iter()
+        .filter(|f| has_ext(&f.path, "txt"))
+        .map(|f| f.path.clone())
+        .collect()
+}
+
+fn has_ext(path: &Path, expected: &str) -> bool {
+    path.extension()
+        .and_then(|e| e.to_str())
+        .map(|e| e.eq_ignore_ascii_case(expected))
+        .unwrap_or(false)
+}
+
+/// Read a text file for harvesting, capped at `MAX_TEXT_FILE_BYTES`, decoding
+/// with the project's encoding detection (`text_encoding`: UTF-8, UTF-16 BOM,
+/// legacy fallback via chardetng). Returns `None` only on an I/O error, which
+/// is logged so silent skips show up in traces. A non-UTF-8 file is decoded,
+/// not dropped.
+fn read_capped_text(path: &Path) -> Option<String> {
+    use std::io::{ErrorKind, Read};
+
+    let f = match std::fs::File::open(path) {
+        Ok(f) => f,
+        Err(e) if e.kind() == ErrorKind::NotFound => {
+            debug!(
+                "candidate_text: file vanished after scan: {}",
+                path.display()
+            );
+            return None;
+        }
+        Err(e) => {
+            warn!("candidate_text: open failed for {}: {e}", path.display());
+            return None;
+        }
+    };
+    let mut buf = Vec::new();
+    if let Err(e) = f.take(MAX_TEXT_FILE_BYTES).read_to_end(&mut buf) {
+        warn!("candidate_text: read failed for {}: {e}", path.display());
+        return None;
+    }
+    Some(crate::text_encoding::decode_text(&buf).text)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn enumerate_filename_inputs_skips_all_audio_and_cue() {
+        use crate::import::folder_scanner::{AudioContent, ScannedCueFlacPair, ScannedFile};
+
+        let cue = ScannedFile::new(
+            PathBuf::from("/rel/Album.cue"),
+            "Album.cue".to_string(),
+            100,
+        );
+        let flac = ScannedFile::new(
+            PathBuf::from("/rel/Album.flac"),
+            "Album.flac".to_string(),
+            5_000_000,
+        );
+        let cover = ScannedFile::new(
+            PathBuf::from("/rel/Artist Name - Album.png"),
+            "Artist Name - Album.png".to_string(),
+            10_000,
+        );
+        let pair = ScannedCueFlacPair {
+            cue_file: cue.clone(),
+            audio_file: flac.clone(),
+            cue_sheet: None,
+            total_size: 5_000_100,
+        };
+        let categorized = CategorizedFiles {
+            audio: AudioContent::CueFlacPairs {
+                pairs: vec![pair],
+                format_label: "CUE+FLAC".to_string(),
+            },
+            artwork: vec![cover.clone()],
+            documents: Vec::new(),
+            unpaired_cue_sheets: Vec::new(),
+        };
+
+        let inputs = enumerate_filename_inputs(&categorized);
+        assert!(
+            !inputs.iter().any(|p| p == &cue.path),
+            "CUE names come from the parsed sheet, not the filename pool; got {inputs:?}",
+        );
+        assert!(
+            !inputs.iter().any(|p| p == &flac.path),
+            "audio filename stems are almost always track titles — wrong pool \
+             for Artist / Album autocomplete; got {inputs:?}",
+        );
+        assert!(
+            inputs.iter().any(|p| p == &cover.path),
+            "artwork filenames still contribute — `Artist Name - Album.png` \
+             and similar carry real signal; got {inputs:?}",
+        );
+    }
+}
