+++ bae-core/src/signals/mod.rs
@@ -31,7 +31,11 @@ pub mod barcode;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
 pub mod disc_id;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
+mod fast_pass;
+#[cfg(not(any(target_os = "ios", target_os = "android")))]
 pub mod origin;
+#[cfg(not(any(target_os = "ios", target_os = "android")))]
+mod pool;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
 pub mod service;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
